// InventarioCliente/ClientApp/src/models/PersonaVM.ts
export class PersonaVM {
  id: number = 0;
  documento: string = "";
  nombreCompleto: string = "";
  primerNombre: string = "";
  segundoNombre: string = "";
  primerApellido: string = "";
  segundoApellido: string = "";
  fechaNacimiento: string = "";
  direccion: string = "";
  telefono: string = "";
  email: string = "";
  tipoDocumento: string = "";
  area: string = "";
  tdocumentoId: number = 0;
  areaId: number = 0;
}